generic_views/detail: let relation fields use a simple widget
Since the detail view is a read-only view, we don't need to fetch the
full queryset of possible "choices" for a related field. The performance
implications of that became apparent when using 5k of ItemTemplates, 2k
of Locations.
We can just query the single object we have as 'value' and get its name
and href.