import_ktim: optimize access to the delegates records for supplier
It has been noted that the procurements.delegate table was queried
too often during the import.

What happened: in order to prepare the 'defaults' for a possible
create(), the delegate was queried a few values. However, these
values were actually in the 'common.partner' table, which meant
that separate queries were issued for each attribute (observed with
Django 1.3). This killed performance.

Now, since our mapping is delegate->supplier, keep a cache of that.