generic_views: enhance the view filters, char field and OR-ed ones
When a user searches, they don't want to bother with huge selection
options or exact strings. They want to input a few letters and let
the DB be searched for that pattern.

So, some Selection fields will be replaced by plain text input ones.
Also, we may want to search in several model fields by the same
pattern. Allow destination to be ('field1__icontains', 'field2__eq') ,
for example.