main.carts: first draft of "cart" functionality
Inspired by the e-shop UI paradigm, created a "cart" data structure
inside the http session. This will be used to let some models "fill"
the carts of others.

We support multiple carts. Hence the user can work in multiple relations
simultaneously.

This is just a draft. Not much of the intended functionality yet.