bloatedListView: put the inner header inside the expanded rowset
We had 2 headers: one for the groupped row and one for the inner, expanded
ones. The second one would be confused with the inner rows, and thus
shall better be used /below/ the group, just above the expanded rows.

Note that this only works under the assumption that we only expand one
group at a time. If we ever want to expand more groups, we have to put
the header back on top.