generic_views: with "generic_layout_form" let the app decide the layout
This puts back some order in the chaos of designing our own views.
In an attempt to avoid copy-pasting long view (template) definitions,
we had created the "generic views", which would use a common template
and just have all their fields in a vertical list.

This is not the UI we really want. We have to explicitly define the
layout of each view, but NOT want repeat the "field" sub-template.

So, rearrange the templates. Put all framework code in one and keep
the "form_layout" as a nice and clean block for each application view.

More to come in a few commits...