all: major consolidation of MEDIA->STATIC files
According to:
    https://docs.djangoproject.com/en/1.4/howto/static-files/

as of Django 1.3+, we are strongly advised to use the "static" paths
of misc. files, rather than the "media" mechanism.
First reason is that MEDIA are supposed to be writable by the user,
through uploads, while STATIC is served statically.

Second reason is that the previous layout of the files was a mess. We
have had "jquery-*.js" in so many places, including it into conflicting
objects. Also, media was dispersed into strange folders in the project
tree.

By this commit, we cleanup the static files as in the Django 1.3+ advice.