products/form_fields: Categories Attributes widget
So, a product (aka. ItemTemplate) may have some set attributes, according
to its category. We want this to appear at a custom widget, where each
attribute will be a combo of possible values. This is a bit awkward to
the MVC structure of Django, since the combos and values depend on the
category of the product.

In implementation terms, the Field+Widget expect a dict as `initial`
value, which will contain a 'from_category' item. Then, all combos will
have the same `name` in html, so that all selections are sent together
by the browser, in a list.