movements: when closing an inbound move, update property numbers
A property number is only assigned once for the lifetime of an asset,
but also depends on the Department the asset was received to. This
means it will /not/ work early, when the asset is created for a draft
movement. So, we have to save() again after the Location is updated,
when the department is set.