movements: do connect the bundled items to their container bundle
Plug a big functional hole in the previous code: after bundles are
created in a move to the Department location and bundled accessories
are created in a move to "Bundled" loc., the accessories must be
linked as assets to their containers.

This algorithm has been written in a "remedy" manner, so that existing
data can be updated, too.