movements: ..but, allow superuser to validate movements
We had restricted the validation action only to Department members with
sufficient privileges. NOT even the "staff" users can validate, because
this action would credit the Department with valued equipment.

Still, the superuser shall have the permission to do that. They do have
access to all other sensitive areas, anyway.