movements.forms: resolve AutoComplete fields to plain combos, if able
AutoCompleteFields had been used for the "location" foreign field,
because the number of locations is great. However, in a parallel motion,
Department users were restricted only to manage their own locations.
So, now it makes sense to resolve those fields to a much simpler combo
box. But only when the request suggests we have a limited set of choices.

This will introduce +2 queries for each such field, during form rendering.
It will, on the other hand, take away the whole AJAX procedure, if the
queryset is small.