generic_views: cleanup the read-only mixin, it's only for widgets!
I had confused, in previous commits, the usage of _RO_mixin, thus
falsely applying it to Fields. It's supposed to be for widgets, and
the bound_data() patch can apply to fields in a way that Field classes
need not be overridden.

Also, renamed the mixin to _ROw_mixin, to remind me it's for widgets