generic_views: more extra_fields for DetailForm
In the DetailForm, we can add some fields with the 'extra_fields'
attribute. These could include the ManyToMany fields, which are not
part of the model.fields .

But, some functionality was missing: We need to have a name different
than the attribute (eg. in cases that the attr. needs to have dots).
In addition, need to specify the widget or get it from form.Meta.widgets .

This time we also support ManyToMany natively, recognizing their
ManyRelatedManager .