login_middleware: after login, redirect to original request page
Suppose we wanted to see "/assets/asset/1" but are not logged-in yet.
This middleware would kick-in and redirect to "/login" . But it's
better to append the "?next=/assets/asset" so that we come back to this
page after login.

With CAS backend, this would be done with the HTTP_REFERER, which is
sometimes wrongly set.