movements: new 'RepairOrder' model
After a repair order is entered in the special view, there are some
movements created, which convey most of the information about the
repair. However, there is still some data that cannot be inferred
by these movements. That is, the /groupping/ of several movements
(and thus the single-point validation) is what we cannot perform
using separate moves. Nor could we list the "repairs" as such.

Now, we introduce one more model, with a ForeignKey at the Movement,
namely the "RepairOrder", to group all the moves of a repair.

Remember to do:
ALTER TABLE movements_movement ADD repair_order_id INTEGER REFERENCES movements_repairorder(id);