po_wizard: don't validate group+bundles without their parts
We now use is_group=True,is_bundle=True for groups, so the 'bundle'
validation algorithm would kick in. But it would only consider the
parts, not the referencing lines to the group. So it would falsely
display the group as "missing", if the group requires at least one
mandatory component.

When loading the po wizard from a stored PO, we just skip the test
and display a blank state. The transition from "4"->"5" will check
again anyway.