generic_views: fix algo of tree_field for non-selectable parents
The tree_field is used to lay Choices in an options group tree.

We also supply the queryset of valid choices + the "parent" field
to group with.

The old algo would unconditionally append the /parent/ choice as
a valid one, even if that weren't included in the supplied queryset.

Now, have a "selectable" flag within the algorithm, so that we
can avoid these Choices in the final layout.

From Ticket: #85