moves/PO-copy: details must be rendered outside of main form
The "details" form for the source PO does contain a subtemplate,
which in turn has a POST form for the paginator. This would
break the DOM of the main form[1] and also falsely submit it
before the user finishes entering the data.

So, we split that part in a special block outside the main form,
but also introduce a "submit=1" parameter to distinguish between
paginator POST or final submission.