movements: at PO confirm, ask movement name, close those first
At PO confirm, we may close our Department's moves, but not any
of other Depts. That's why the "bundle" move has to be confirmed
last. If there is no "name" in a move, we must ask the user again.

Also correct the previous logic for showing the confirmation
question. It didn't work for super-users.