inventory/views: allow superuser to validate, in Development mode
At a production database, nobody but the "validator" user shall be
allowed to validate an inventory. Not even an administrator, because
this would register a person not legaly responsible for that action.

But, in development, we can't work that way. A superuser won't have
an active role, neither fine-grained access permissions.