movements: lock PurchaseOrders against double validation
We have witnessed a few cases where users are impatient and
click on PO validation twice. Then, long processing times for
movements+items may also allow a window for double-processing
of that data.

Locking the row in SQL wouldn't do, because we don't want to
wait. SELECTing rows that don't exist, doesn't either.
Instead, we temporarily change the state, let other threads
know that PO is in process, keep an object that will revert
that state in any case, when deleted.