movements/mgmt: report excess+missing items, if these mismatch
In an interesting twist, some user had edited a serial number
after the PO was received. Our algorithm could not match these
two lines and reported the Item as being excess. However, one
similar item was missing, of the same product.
In such a case, the admin must act manually.