reports/app: workaround for multiple POST of results
When clicking the results button, we do a POST in order to send
all the report configuration up to the server for rendering.
However, browsers don't allow a second POST to the same URL,
from the same form, because POST methods are considered non-safe.

We trick the browser into submitting again with a counter that
increases at each form submission. We also send the report id,
if available, so that the new page can be refreshed.

Unfortunately, angularJS wouldn't allow us to bind these vars.
directly into the "action" attribute, because it cannot trust
a mix of static and bound strings. So we manipulate the form
action through plain JS accessors.