reports: workaround Django failing query of inherited models
We have, say, two models: Manufacturer inheriting Partner and
products.ItemTemplate referencing its "manufacturer"

If we try to query:
    ItemTemplate.objects.order_by('manufacturer_id').values(...)
the query borks, because it JOINs Partner but not Manufacturer.
So, we have to order by natural:
    ItemTemplate.objects.order_by('manufacturer').values(...)
Which works.

But, then, we still need to look for "manufacturer_id" in our
values, to match against.