reports/views: "extra query" support, for non-trivial subqueries
For quantitative constraints, we cannot just use some JOIN on
the other table. We need, instead, a complete sub-query, that
computes an aggregate (SUM or COUNT) and then apply a clause
on that aggregate.

Django can only support that through `.extra()` query clauses.

Unfortunately, we cannot put the sub-query in a "select" field,
because these fields will be only computed on demand (depending
on the outer-query fields) and certainly not participate in the
aggregate queries. So, we put the whole sub-query inside "where"
extra-clause.

This is the domain implementation for 32ca8582b370dac (2/3)