routing: query multiple databases, with smart routing
Django allows to have multiple database connection settings,
from which the proper DB is selected for each operation.
We introduce a "smart" router, that will have clusters of
such settings and pick a DB from the cluster each time.

This way we can load-balance our read operations, or even
direct reporting queries to a different set of servers.