reports: substitute parameters in domain expressions
Although the "domain" for reports may come from the client
side, there is still a case that we want extra parameters
to the report:
 users may be granted explicit access to some saved reports,
but still need to specify some criteria.

So, let (string) fields be substituted. eg. the domain can
be like:
   [('name', '=', 'some $foo')]
and then send parameters `foo=thing`, so that query becomes
   [('name', '=', 'some thing')]

Admin accounts (only) see this in the reports UI, all users
may use these reports.