movements/wizard: use autocomplete.combobox for long lists
Manufacturers and the final product combo may take quite many
candidates. The default html drop-down selection isn't usable
in that case.
Fortunately, jQuery-UI provides the "combobox()" as an example
for drop-down + text filter.
I only had to fix the "select" event to trigger `change`, as
we heavily depend on that.