all: prepare site for ajax-select introduction
At grappelli, which overrides the admin base template, we need to open
a new block (aka. extraheaders2) because the order of inheritance does
not allow us to fill the "extraheaders".

The ajax_select.js and .css are centrally added to the base templates,
for efficiency.
jQuery-ui is now included.
django.contrib.staticfiles is included, in order to serve ajax_select.

Conflicts:
	settings.py