common/api: truncate whitespace in Lookup responses
Data from lookup channels will be rendered as HTML, where
whitespace is all truncated to a single space, by default.
Now, apply that logic in ajax response data, so that it can
be sorted regardless of whitespace.

This fixes the case where product attributes are combined
into their names with non-visible whitespace characters.