movements,inventory: enforce SQL transactions on critical actions
After discovering inconsistencies, we want to strengthen the logic
of transactions.
By default, Django 1.4 [1] will open a transaction, but commit early
after each .save() . This means that eg. closing a movement will be
performed in 2-3 separate commit()s, instead of a single transaction.

Now, we try to keep the transaction open and uncommitted until the
procedure comes to a successful end. Manual `rollback()` is needed in
error paths (because our code does render an error page instead of
raising the exception higher).
Then, we need to force SQL row-level locking from the time we "touch"
some critical lines (eg. state of Items, Movements) until they have
been UPDATEd. This is done with "select_for_update()" decorator on
relevant querysets.

This may still not be complete, and may even cause contention of user
requests waiting for those locked rows. But is better than just early
commits in the middle of these processes.


[1] as of 1.6, this is even worse, as autocommit is chosen by default.