movements/cmd/verify: use manual transactions for each batch of items
Technically, each management command would be a single SQL transaction.
But, with autocommit turned on that means each ".save()" would trigger
a commit. This is marginally safe for "select for update", since save()
only happens at the end of each (atomic) asset calculation.


Still, this means that select for update may hold locks for quite long,
unless there is some need to "save()" things.

Now, do an explicit transaction, hold the update lock for the duration
of each batch.