movements/models: when draft moves removed/rejected, reset bundles
During the PO wizard, bundling in newly-created ItemGroups is computed
and stored /before/ the moves are confirmed. This is in order to have
the bundle state appear early to the user.

However, if any of those moves is deleted or rejected, bundles are not
cleared accordingly, resulting in false display of parts in them.

Also, have a step in "movements_verify" management command to clear
these cases.