movements/po_wizard: update moves when PO.issue_date changes
Currently there is no UI to modify draft movements. Many users have
been re-visiting "step 1" of the wizard in order to update the PO's
issue_date. Created movements (after step 4) would not be updated,
before this commit.

We try to be smart and NOT update, however, movements that are done
or have a different initial date, anyway.